Trying to be better about error handling

+++ kernel/include/libc/string.h
@@ -9,8 +9,8 @@ void * memcpy(void * dest, const void * src, size_t n);
 void * memmove(void * dest, const void * src, size_t n);
 void * memset(void * dest, uint8_t value, size_t n);
 
-size_t strlen(const char * str);
-size_t nstrlen(const char * str, size_t max);
+int strlen(const char * str);
+int nstrlen(const char * str, int max);
 int strfind(const char * str, size_t start, char c);
 char * strtok(char * str, char * delim);
